Cluster addition implemented
