+++ .gitignore
@@ -1,2 +1,4 @@
 venv/
 __pycache__/
+*.pyc
+*.db
