+++ class_doc/kmeans/2009pivot.csv
