+++ class_doc/kmeans/st99_d00.dbf
