+++ class_doc/spectral/spectral01.py
