+++ class_doc/spectral/spectral02.py
